add thread/process priority option
