+++ autd3capi-link-soem/Cargo.toml
@@ -16,3 +16,4 @@ static = []
 autd3capi-driver = { workspace = true }
 autd3-link-soem = { workspace = true, features = ["local", "remote", "async-trait"] }
 tracing-subscriber = { workspace = true }
+thread-priority = "1.1.0"
